backend/requirements.txt
# FastAPI and Server
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.1.1
pydantic==2.11.9
pydantic-settings==2.11.0
python-multipart==0.0.20

# Database
psycopg2-binary==2.9.10
sqlalchemy>=2.0.25
pgvector==0.4.1

# Twilio WhatsApp Integration
twilio==9.8.3

# OpenAI & LangChain
openai==2.2.0
langchain==0.3.27
langchain-openai==0.3.35
langchain-community==0.3.30
langchain-core==0.3.78
langchain-text-splitters==0.3.11

# PDF Processing
pypdf==6.1.1

# Vector Store
faiss-cpu==1.12.0
